Moved to article recommendation to opensearch module
